User logs in via MapMyFitness
Sets up the basic oauth2 using an omniauth map-my-fitness gem. Creates the
User table in the database and the find_or_create_by_auth method for the user
which allows a user to be looked up via the provider and their user id for that
provider. Sets up super basic dashboard and welcome views.